An application that use huffman encoding to compress data structures and allows efficient traversal
Huffman coding algorithm is a type of optimal prefix code, which is extensively used for lossless data compression. Normally, the data which can be viewed as a collection of symbols (which usually includes alphanumeric characters, especially
in our case) is stored in ASCII code which uses 8 bits per symbol memory space. But using Huffman coding, variable length codes can be generated in place of the
standard ASCII code to store the data, hence resulting in reduced memory storage. This algorithm is applicable whenever memory efficiency is indispensable.